add listing base class
