+++ starfish/listing/__init__.py
@@ -1,2 +1,3 @@
 
+from starfish.listing.listing_base import ListingBase
 from starfish.listing.listing import Listing
